+++ control_utils.py
@@ -5,21 +5,34 @@ from .db_utils import DBUtils
 from .docker_utils import DockerUtils
 from sqlalchemy.sql import and_
 from flask import session
-
+from .redis_utils import RedisUtils
 
 class ControlUtil:
     @staticmethod
-    def add_container(user_id, challenge_id, flag, port=0):
+    def add_container(app, user_id, challenge_id, flag, port=0):
         uuid_code = DBUtils.create_new_container(user_id, challenge_id, flag, port)
-        DockerUtils.add_new_docker_container(user_id=user_id, challenge_id=challenge_id, flag=flag, uuid_code=uuid_code)
+        DockerUtils.add_new_docker_container(app=app, user_id=user_id, challenge_id=challenge_id, flag=flag, uuid_code=uuid_code)
 
     @staticmethod
-    def remove_container(user_id):
-        docker_result = DockerUtils.remove_current_docker_container(user_id)
+    def remove_container(app, user_id):
+        docker_result = DockerUtils.remove_current_docker_container(app, user_id)
         if docker_result:
+            container = ControlUtil.get_container(user_id)
+            port = container.port
             DBUtils.remove_current_container(user_id)
+            if port != 0:
+                redis_util = RedisUtils(app)
+                redis_util.add_available_port(port)
 
         return docker_result
+
+    @staticmethod
+    def get_container(user_id):
+        return DBUtils.get_current_containers(user_id=user_id)
+
+    @staticmethod
+    def renew_container(user_id, challenge_id):
+        DBUtils.renew_current_container(user_id=user_id, challenge_id=challenge_id)
 
     @staticmethod
     def check_challenge(challenge_id, user_id):